Refactor the project entirely
1. Remove the previous redundant Node object, since the instantination of such objects are not necessary.
2. Implemented the heuristic evaluator with two 'splitting 1024 tile' strategies.
3. Reserve interfaces for dll exporting.

+++ engine/include/HeuristicEvaluator.hpp
@@ -9,7 +9,7 @@ class HeuristicEvaluator : public IEvaluator {
 public:
     void initialize() const;
 
-    double evaluate(board_t board) const;
+    float evaluate(board_t board) const;
 };
 
 } // namespace game2048
